Cleanup: Move /config.json to /config/config.json so that it matches up to where OpenShift can mount the config file.  This avoids needing to setup a symlink.
