+++ app/ui/.dockerignore
@@ -1,1 +1,1 @@
-dist/config.json
+dist/config/config.json
