+++ app/ui/angular.json
@@ -26,7 +26,7 @@
             },
             "assets": [
               {
-                "glob": "config.json",
+                "glob": "config/config.json",
                 "input": "src",
                 "output": "/"
               },
