+++ app/ui/scripts/generate-ts.js
@@ -20,7 +20,7 @@ const fs = require("fs");
 const argv = require("process").argv;
 const env = require("process").env;
 const url = require("url");
-const config = require("../src/config.json");
+const config = require("../src/config/config.json");
 const CodeGen = require("swagger-js-codegen").CodeGen;
 const pluralize = require("pluralize");
 const templates = path.join(__dirname, "templates");
