+++ app/ui/src/config/config.json.minishift
